fixed a few warnings
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/anatomist/anatomist-free/branches/bug_fix@75300 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/lib/control/wPreferences.cc
@@ -495,14 +495,14 @@ PreferencesWindow::PreferencesWindow()
       Object oi = oc->objectIterator();
       if( oi->isValid() )
       {
-        col[0] = oi->currentValue()->getScalar() * 255.99;
+        col[0] = uint8_t( oi->currentValue()->getScalar() * 255.99 );
         oi->next();
         if( oi->isValid() )
         {
-          col[1] = oi->currentValue()->getScalar() * 255.99;
+          col[1] = uint8_t( oi->currentValue()->getScalar() * 255.99 );
           oi->next();
           if( oi->isValid() )
-            col[2] = oi->currentValue()->getScalar() * 255.99;
+            col[2] = uint8_t( oi->currentValue()->getScalar() * 255.99 );
         }
       }
     }
@@ -1300,14 +1300,14 @@ void PreferencesWindow::changeWindowBackground()
       Object oi = oc->objectIterator();
       if( oi->isValid() )
       {
-        col.setRed( oi->currentValue()->getScalar() * 255.99 );
+        col.setRed( uint8_t( oi->currentValue()->getScalar() * 255.99 ) );
         oi->next();
         if( oi->isValid() )
         {
-          col.setGreen( oi->currentValue()->getScalar() * 255.99 );
+          col.setGreen( uint8_t( oi->currentValue()->getScalar() * 255.99 ) );
           oi->next();
           if( oi->isValid() )
-            col.setBlue( oi->currentValue()->getScalar() * 255.99 );
+            col.setBlue( uint8_t( oi->currentValue()->getScalar() * 255.99 ) );
         }
       }
     }
